Renamed package name from com.examples to com.deutschlernen.
 * Made changes to reflex new package name in all the files.
 * Added custom app icons in Manifest
