+++ app/src/main/AndroidManifest.xml
@@ -1,16 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.willkommenistdeutsch">
+    package="com.android.deutschlernenzuzammen">
 
     <application
         android:allowBackup="true"
-        android:icon="@mipmap/ic_launcher"
+        android:icon="@mipmap/ic_launcher_round"
         android:label="@string/app_name_short"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/DarkTheme">
-        <activity android:name=".AboutUsActivity"></activity>
-        <activity android:name=".MainActivity">
+        <activity android:name="com.deutschlernen.deutschlernenzuzammen.AboutUsActivity"></activity>
+        <activity android:name="com.deutschlernen.deutschlernenzuzammen.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
