+++ app/src/main/java/com/deutschlernen/deutschlernenzuzammen/fragments/TrennbarUntrennbarFragment.java
@@ -1,6 +1,4 @@
-package com.example.willkommenistdeutsch.fragments;
-
-import androidx.fragment.app.Fragment;
+package com.deutschlernen.deutschlernenzuzammen.fragments;
 
 import android.content.Context;
 import android.os.Bundle;
@@ -8,9 +6,12 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.example.willkommenistdeutsch.R;
+import androidx.fragment.app.Fragment;
 
-public class WeekNamesFragment extends Fragment {
+import com.android.deutschlernenzuzammen.R;
+
+
+public class TrennbarUntrennbarFragment extends Fragment {
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -18,9 +19,10 @@ public class WeekNamesFragment extends Fragment {
     }
 
     @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState)
-    {
-        return inflater.inflate(R.layout.fragment_week_names, container, false);
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.trennbar_untrennbar, container, false);
     }
 
     @Override
@@ -32,4 +34,5 @@ public class WeekNamesFragment extends Fragment {
     public void onDetach() {
         super.onDetach();
     }
+
 }
