+++ app/src/main/java/com/deutschlernen/deutschlernenzuzammen/fragments/WeekNamesFragment.java
@@ -1,4 +1,4 @@
-package com.example.willkommenistdeutsch.fragments;
+package com.deutschlernen.deutschlernenzuzammen.fragments;
 
 import androidx.fragment.app.Fragment;
 
@@ -8,9 +8,10 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.example.willkommenistdeutsch.R;
+import com.android.deutschlernenzuzammen.R;
 
-public class GreetingFragment extends Fragment {
+
+public class WeekNamesFragment extends Fragment {
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -20,7 +21,7 @@ public class GreetingFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState)
     {
-        return inflater.inflate(R.layout.fragment_greeting, container, false);
+        return inflater.inflate(R.layout.fragment_week_names, container, false);
     }
 
     @Override
