+++ app/src/main/java/com/deutschlernen/deutschlernenzuzammen/fragments/WhQuestionFragment.java
@@ -1,4 +1,4 @@
-package com.example.willkommenistdeutsch.fragments;
+package com.deutschlernen.deutschlernenzuzammen.fragments;
 
 import android.content.Context;
 import android.os.Bundle;
@@ -8,9 +8,8 @@ import android.view.ViewGroup;
 
 import androidx.fragment.app.Fragment;
 
-import com.example.willkommenistdeutsch.R;
+import com.android.deutschlernenzuzammen.R;
 
-import java.util.zip.Inflater;
 
 public class WhQuestionFragment extends Fragment {
