+++ app/src/main/res/layout/activity_about_us.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:background="?attr/backgroundcolor"
-    tools:context=".AboutUsActivity">
+    tools:context="com.deutschlernen.deutschlernenzuzammen.AboutUsActivity">
 
     <TextView
         android:layout_width="match_parent"
