+++ app/src/main/res/layout/fragment_adjektiv_endings.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.AdjektivEndingsFragment"
+    tools:context="com.deutschlernen.deutschlernenzuzammen.fragments.AdjektivEndingsFragment"
     android:background="?attr/backgroundcolor"
     >
