+++ app/src/main/res/layout/fragment_dativ_verben.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.DativVerbenFragment"
+    tools:context="com.deutschlernen.deutschlernenzuzammen.fragments.DativVerbenFragment"
     android:background="?attr/backgroundcolor"
     >
