+++ app/src/main/res/layout/fragment_greeting.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.GreetingFragment">
+    tools:context="com.deutschlernen.deutschlernenzuzammen.fragments.GreetingFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
