+++ app/src/main/res/layout/fragment_rules.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.RulesFragment"
+    tools:context="com.deutschlernen.deutschlernenzuzammen.fragments.RulesFragment"
     android:background="?attr/backgroundcolor">
 
     <LinearLayout
