+++ app/src/main/res/layout/trennbar_untrennbar.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.TrennbarUntrennbarFragment"
+    tools:context="com.deutschlernen.deutschlernenzuzammen.fragments.TrennbarUntrennbarFragment"
     android:background="?attr/backgroundcolor">
 
     <LinearLayout
