+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 <resources>
     <string name="app_name">Herzlichen Willkommen.!\n\nFangen Wir Deutsch lernen an.!</string>
-    <string name="app_name_short">Fangen Wir Deutsch lernen an.!</string>
+    <string name="app_name_short">Deutsch Lernen Zuzammen</string>
     <string name="about_us">About Us</string>
 
     <string name="developer_name">Shivam Naik</string>
